requirements.txt
appdirs==1.4.4
asgiref==3.4.1
bidict==0.21.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
clipboard==0.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
django-socketio==0.2
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.31.1
filelock==3.0.12
gevent==21.1.2
gevent-socketio==0.2.1
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
pipenv==2021.5.29
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.15.0
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
sphinx-me==0.3
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
waitress==2.0.0
webdriver-manager==3.4.2
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.4.0
